backend/requirements.txt
APScheduler==3.10.1
Django==4.2.3
django-apscheduler==0.6.2
djangorestframework==3.14.0
requests==2.31.0
mysqlclient==2.2.0
python-decouple==3.6
python-telegram-bot==21.4